TPA: A note for future optimization
If TPA found transition invariant restricted to init, instead of
returning the negation of init as the explanation of safety
(over-approximation of initial states that are known to be safe). we
could compute the safe inductive invariant.
This is also an over-approximation of initial states that is guaranteed
to not reach bad states. It does require quantifier elimination, which
could be expensive. There could be a time limit on the computation with
the fallback to just the initial states.