TermUtils: Simplify TrivialQuantifierElimination
For large number of variables, the original implementation was repeating
a lot of work: Collecting top-level equalities and substituting for
each variables.

New version builds a single substitution map and does the simplification
only once. However, to deal with substitution cycles, we now simply
forbid any variables used in a definition of other to have definition
themselves. This means we may not eliminate as many variables as before.
But the implementation is much faster than before. Though it still may
be improved.