Normalizer: Support multiple instances of the same predicate in clause's body
Previously, in nonlinear systems, if there were multiple instances of
the same predicate in the body of the clause, normalizer would use the
same variables for all instances, which is not wrong.

Now, we create different version of variables. We still need to see how
engines supporting hypergraphs (only Spacer at the moment) should
deal with this situation.