Everywhere: Fix handling of nonlinear CHC systems
Previously we did handle nonlinear CHC sytems correctly, when a clause
with multiple instances of the same predicate were present.

The normalization process was fixed before but the previous version was
difficuly to handle in the Spacer algorithm.
Here we introduce a different way of handling the problem: We introduce
source and target tags on the variables (source for predicates in the
body, target for the predicate in the head). Additionally, the source
variables are versioned based on the number of occurences of the given
predicate. We also introduce VersioningManager to handle translation
between base, source and target versions of variables and formulas.
This is enough to represent the normalized version of the system, as
well as for algorithms that work with single copy of the transition
relation, such as Spacer.
For compatibility, we keep the previous versioning system with
TimeMachine for linear CHC systems, but we need to introduce a
translation between the different canonical representation of variables
in linear and nonlinear systems.

TODO: We need to find a way to unify the handling of linear and
nonlinear systems. Step versioning (now used for linear systems) will be
needed also for nonlinear systems to represent counterexamples.