Interpreter: Enable simple reduction on ChcGraph
The reduction contracts all such nodes in the graph that has single
outgoing or single incoming edge. These nodes are redundant, contracting
them does reduces the number of edges in the graph (by one).

TODO: We need to adjust the solutions of the modified graph to get
solution for the original graph.