LAWI: Fix bug in handling coverings for error locations
Previously, when a label of an error location was strengthened, we did
not update the covering relation for such vertex. This situation is
unusual, but can happen, e.g., when there are multiple edges from some
location to the error location. In this case, an error node in ART
may be covered by another error node not yet explored.
When such node is finally explored, its label is strengthened and we
must remove it as a coverer from covering location. Otherwise, the
algorithm would incorrectly conclude the graph safe.