Everywhere: More flexible CHC graph representation
Using a vector of vertices with ID as index to this vector is
impracticle in situations where we need to modify the graph, which every
preprocessing pass will do.
Next attempt is to use predicate symbols directly to represent vertices,
and store only edges (+ predicate representation) to represent the
graph.
This should provide more flexibility for graph modification, as well as
make it easier to generate solution back-translators from pre-processing
passes.