Transform: Fix detection of trivial node
Previously, even if node was a target of a hyperedge, it could be
considered trivial. However, when we are trying to detect chains of
trivial node, the need for special handling of first edge would
complicate the code unnecesarily. Instead, we are gonna require that
both the single incoming and the single outgoing edges do not have more
than one source.