Spacer: Optimize SpacerContext::mayReachable
Previously, SpacerContext::mayReachable, which computes if a target
constraint can be reached using a given edge, was computing an
interpolant in case the answer was UNREACHABLE.
However, the interpolant was never used.

Here we change the method so that it returns only YES/NO answer. The
benefit is that the underlying check does not need to enable interpolant
computation, so it can be faster. A quick check revealed that
computation of substitutions in these checks now might be a bottleneck.