Transformers: Implement backtranslation of proofs in SimpleChainSummarizer
This implementation is a prototype and a little bit cumbersome.
The goal is to replace a single derivation step by a sequence of
derivation steps that correspond to the summarized chain.
Currently we proceed one summarized step at a time, and patching the
proof after each replacement.
The pathc is required because indices must correspond to the position of
the steps and we are shifting all the steps after the replaced one.

This is an inevitable consequence of the current design.
It might pay of to investigate other possible designs.