Spacer: Fix bug in handling 0-ary predicates in witness computation
When computing premise instances from constraints on variables in the
database, we previously tried to use a shortcut and just evaluate the
predicate in the model of the constraints.
However, this does not work correctly for 0-ary predicates, since Model
class will try to evaluate it as an unknown boolean variable and just
return "true" for it.
This is incorrent, the instance should be the predicate itself.

Here we fix it by properly obtaining the values for all the variables
from the model and then just creating the instance by the usual
mechanism of Logic.