Normalizer: Create less terms in the normalization process
Number of terms in Logic currently has an impact on the solving process,
especially in case we are asking a huge amount of simple queries, e.g.
in the Spacer engine.
When the queries are simple, the runtime is dominated by setting up (and
tearing down) the solvers.
The problem is that in OpenSMT, we are using some data structures that
use Pterm's id as an index in an array. Thus, it needs to allocate an
array with size the number of terms in Logic.
Until we change that in OpenSMT, we can remedy some of the problem by
creating less terms in Golem.

Profiling showed that a lot of these terms are created in the
normalization process, especially with systems of many clauses with many
variables.
During normalization, we create three versions of variables for each
predicate: versionless, source and target.
Then we also create equalities that relate normalized variables to
original arguments.
However, if the original arguments were variables as well, these
equalities are only temporary. Instead of creating these temporary
equalities, we can directly apply substitution, thus avoiding the
creation of (possible large amount of) temporary terms.