Spacer: Do not repeat work when checking edges with single source
When checking reachability of a proof obligation, we first check
reachability with must summaries and then unreachability with may
summaries. If both checks fails, then we systematically iterate over
vertices to find the first one which makes reachability feasible when
may summary is used instead of the must summary.
For edges with single source this unnecessarily duplicates work, because
when we fail to prove unreachability with may summaries, we already have
the model to generate proof obligation at that point, so we do not need
to repeat the same check later.