TPA: Add functionality for solving trivial graphs
This is a quick fix TPA to be able to handle instances with no loops.
Since TPA eliminates non-looping nodes, the trivial graph will have (0
or more) edges from entry to exit.

This should be probably generalized. Nonrecursive CHC systems can be
solved prior to passing the problem to the engines.
One way is to solve it as a single tree-interpolation problem.
An easier way (with less new code) is to apply non-loop-node elimation
to obtain a trivial graph and let the backtranslation compute the
witness.