Transformations: Implement canonical transformation Lin-CHC -> TS
For each predicate, there is a boolean variable that captures if we are
currently at this predicate. Moreover, the transition system has a
variable for each variable in every predicate. Then each clause is
translated to a disjunct of the transition relation, which is enabled
only if the source predicate is enabled (all others are disabled).

We should try using a single integer variable for the predicates,
instead of a boolean variable for each one.