Trans: Proper treatment of auxiliary variables
When translating CHC system to transition system, we need the
information about auxiliary variables as well. Fortunately, we already
have functionality to extract state, next-state, and auxiliary variables
from edge's constraint, so we can just use that!

Note that the correctness depends on the assumption that all auxiliary
variables are unique. This is currently guaranteed by our normalization
process.