Engines: Refactor computation of witnesses for transition systems
This refactoring prepares the code for transformation of general linear
CHC system to a transition system. The plan is that there will be no
CHC graph corresponding to a transition system, but we go to directly to
our inner TS representation. For this reason, we refactor the methods so
they do not require both representation of a transition system.

To support witness production, we add a simple wrapper for a witness for
our inner TS representation.