Transformers: Safety witness translation for single loop transformation
We translate the invariant of the transition system to interpretation of
the predicates restricting the invariant to each location separately.
We do this by setting the current vertex's variable to true and all
other location variables to false.
This can still leave variables from other locations in the restricted
invariant, so we try to find (by simplification) an equivalent formula
over only the variables of the current vertex.
If we are not able to find such a formula, we bail out without the
witness for the linear CHC system.
It would be good to check if there is a better mechanism for this. (Maybe
quantifying the unwanted variables would work?)