IMC: Fix generation of safe inductive invariants
IMC can prove a system safe by discovering an inductive invariant of the
system that is unsatisfiable with SUFF_0^k, that is, it cannot reach bad
states in up to k steps. However, using SUFF_0^k often incurrs overhead
for the solver, so we use SUFF_k^k instead, that is, we only check if
the last state of the suffix is bad. We can still correctly terminate
when we find a fix point in this setting, since we know the bad states
are unreachable; however, the inductive invariant itself is not
necessarily safe, we only know that it cannot reach bad states in k
steps.
In fact, the system added here as a test case witnesses such behaviour.
What we know for certain is that Inv and ~Bad (that is, the intersection
of the discovered invariant with bad state excluded) is a safe
k-inductive invariant. It is safe by definition, and it is k-inductive
because after k-steps we know we still stay in Inv (because Inv is
1-inductive), and we know we cannot reach Bad after k steps when we
start in Inv.

There may be a better way to compute a safe inductive invariant, but
this is the easiest one to implement at the moment.