Transformers: Implement backtranslation for MultiEdgeMerger
For validity witness we do not have to do anything.
For invalidity witness, we can keep the derivation, but we need to
update the metadata, i.e., which edge was used to derive the fact.
This can be done by iterating over the replaced edges and checking which
edge is satisfiable given the values in the premise and conclusion of
the step using the merged edge.