Transformers: Fix backtranslator of MultiEdgeMerger
During backtranslation of the proof, when we search for an original edge
to use instead of the merged edge, we cannot use OpenSMT's model,
because that implicitly assignes default values to unknown variables.
For auxiliary variables possibly present in the labels, this is
undesired behaviour. Instead, we must substitute only the predicate
variables with the corresponding values from the proof step.