TPA: Fix prematurely declared safe inductive transition invariant
When checkExactFixedPoint finds a fixed point for the latest power, we
know the corresponding formula (ATr<n or (ATr<n o ATr=n) is a
2^n-inductive transition invariant, but we do not know if it is safe.
In the next iteration, we check reachability in <2^(n+1) steps, and if
we prove unreachability, we claimed that the previously found formula is
a safe invariant. However, during the reachability check, the
abstraction may have been refined and it may no longer be
(2^n)-inductive.
Here. we check for the fixed point only on the levels where we know that
the correspoding inductive invariants are also safe.