Everywhere: Simplify typical use of SMT solver
The goal is to provide a simple abstraction over OpenSMT's solver so
that we do not have to repeat unnecessary complicated setup of OpenSMT's
SMTConfig and MainSolver. It would also better contain future changes in
OpenSMT.

In the future, this abstraction can be extended to completly isolate
OpenSMT's solver in a single place.