SimpleNodeEliminator: Allow eliminating nodes in hyperedges
This change extends SimpleNodeEliminator so it also eliminates nodes
with single incoming simple edge and any number of outgoing
(hyper)edges. Previously, a node was not eliminated if it participated
in a hyperedge.

The proposed change leads to more nodes eliminated in some benchmarks of
CHC-COMP, which helps Spacer engine.
Additionally, some nonlinear systems can be translated to linear ones by
this transformation, enabling other engines to solve some nonlinear
benchmarks.