Fix incorrect behaviour of MBP
When dealing with divisibility constraints, we need to compute an
extended model where we add a value for a temporary variable.

Previously, the way we computed the extended model was not correct. In a
pathological case, where the implicant contains a component represented
as ~false, the way we tried to compute all the relevant variables would
fail. The reason is that we tried to compute a conjunction of the atoms
(ignoring the polarity) to save time and visit the DAG only ones.
However, in this pathological case, the conjunction would be simplified
to false, and we would not find any variables for building the extended
model.