Transformers: Ensure that auxiliary variables are versioned
At the moment, the codebase assumes that auxiliary variables are
versioned. After the recent changes in the normalization process, we
ended up NOT versioning the auxiliary variables introduced by ITE and
DIV/MOD elimination. This commit fixes this in a bit ad-hoc way.

We should revisit the normalization process and also the invariant about
auxiliary variables being versioned. The presence of both TimeMachine
and VersioningManager makes it difficult to reason about the code's
invariants.