Transformers: Fix backtranslation for RemoveUnreachableNodes
Previosly, the backtranslator for validity witness was adding
interpretation for removed nodes using the source version of the
predicate. However, the rest of the code now works with base versions of
predicates.

Note: We should consider using only predicate name as the key, with the
interpretation still using base version of the formula.