Term: Enable turning raw pointer to shared pointer
Since we build Terms always with shared pointers, this is safe. (Note
that to make this absolutely safe, we should make all constructors
private and use factory to produce instances directly as shared
pointers).

This is useful because in the visitors we only have raw pointers and we
don't have access to the owning shared pointer. This is the safe way to
create shared pointer to Term from a raw pointer.