IMC: Use incremental SMT solving
Instead of moving the intial states through one step forward on each
step, we can keep initial states on version zero.
Then the k-1 transitions (and the query) remain the same in every
iteration, so we can keep them in the solver between iterations.
What changes are the initial states (they grow larger on every
iteration), so we push and pop those in every iteration. To keep things
simple, we push and pop the whole prefix (init + one step), although
even the one step of the prefix could be kept in the solver. Setting the
interpolation mask properly would be a bit more difficult in that case.

With this setup, we can use incremental capabilities of OpenSMT.