Alethe: A step towards supporting non-normalized input
Current Alethe proof production assumes the input system is
normalized in the sense that all arguments of all predicates are
variables.
First step towards supporting general input is that we support general
terms when computing instantiation. For now, we basically ignore general
terms from predicates, and we assume that all variables that are present
in some clause will be some argument of some predicate in the clause.

This fixes the instantiation, but the generated proofs are still not
valid. We are missing steps to proof that the instantiated general terms
are equal to the computed constant value.