Alethe: Fix proof building for non-normalized systems
Previously, if the system contained general terms as predicate arguments
(not only variables), the proof would not be correctly constructed.

Following the strategy outlined in the previous commit, when we derive
RHS of the clause for a given skeleton step, we must further simplify
its arguments if the head of the clause had a general term as one of its
arguments. This is what we add here, the steps to prove that we can
derive the simplified RHS from the general RHS.