Spacer: Avoid creating unnecessary terms
Since OpenSMT is performance-sensitive with respect to the number of
terms created in Logic, we need to try to avoid creating unnecessary
terms.

Specifically, when creating local satisfiability queries, we were
previously encoding that as validity checks, if current lemmas imply the
negation of proof obligation.
However, internally, that was encoded again as a satisfiability check,
so the consequent was negated again.
We can avoid the negation altogether and just do satisfiability checks
directly instead of validity checks.