Spacer: Avoid creating unnecessary terms on proof production
When we need to figure out concrete instances of premises from
constraints and concrete instance of the derived fact, we do not need to
create equalities for fixing the variables to concrete values.
Instead of creating these equalities and passing them to the solver, we
can directly simplify the body and give the solver already a simplified
problem.