TermUtils: Print negative numbers in accordance with SMT-LIB
Unfortunately, we cannot just print the number using Logic::printSym,
because this method is not overriden in ArithLogic, unlike
Logic::printTerm.
For now, we simply work around this by calling printTerm instead of
printSym when dealing with numbers.

It would be better if OpenSMT did this directly by overriding printSym
in ArithLogic as well.