Fix the process of creating CHC representation from input terms
Previously, when extracting uninterpreted predicates from the term
representation, we would not preserve the order of predicates, if
multiple UP were present. This was due to the use of `partition`
algorithm. To preserve the relative order, we must use `stable_partition`.
The produced proofs should now be correct with respect to the provided
input.