IMC: Use only post-image as the next init states
In the inner loop, IMC computes forward image of reachable states by
interpolating over one step of transition relation from the current
approximation of reachable states (considered as the initial states in
the current iteration).
The post-image is then added to the set of reachable states.

However, as McMillan notes is the section Optimizations, there is no
need to compute forward image of all reachable states, but it is
sufficient to consider only the newly discovered states.
Using the whole overaproximation in every iteration repeats a lot of
work unnecessrily.
The interpolant is a good overapproximation of the new states. It still
may contain some already discovered states, but if we use the
interpolant as the new initial states in the next iteration, we do not
need to do any formula manipulation and we still exclude a lot of
already examined states.

Experimental evaluation shows this change has a huge positive effect on
the efficiency of the algorithm.