Graph: Fix computing outgoing edges in adjacency representation
Previously, for a hyperedge where the same node occurs multiple times as
a source, the edge would be added the corresponding number of times as
an outgoing edge for this node.
We now check if the node already has this edge added and do not add it
more than once.