PA: Fix unsoundness caused by incorrect refinement
When some node is refined, we must re-check also its children, because
some of the existing predicates might start being true in them.
When we refine a node, we make its label stronger (label is the
conjunction of predicates that are true at that state).
Thus, some predicates might now be valid in the children.