PA: Check feasibility of an edge just before we add it to ARG
It may happen that an edge is feasible when we add it to the queue, but
later, thanks to the refinement, it may become infeasible before its
turn. In such a case, we would use it to create new nodes when we fetch
it from the queue.  By moving the check to the later stage, we can
preven the creation of redundant nodes in the ARG.