Transformations: First version of edge inliner
Edge inliner checks for each edge if there is at most one feasible
predecessor.
If no predecessor is feasible, the edge can be removed.
If a single predecessor is feasible, the the edge can be merged with the
predecessor. We can remove the original edge but must keep the
predecessor, because there might be other outgoing edges from the
original edge's source.

We also do not want to inline if the predecessor is a loop.
That would basically meant peeling off one iteration from the loop, and that
could continue for a long time.