Graph: Fix computation of auxiliary variables in edge's label
Previously, we detected auxiliary variables in an edge's label simply by
checking the names of the variables.
However, this is not reliable, because preprocessing passes can create
new edges where previously state variables are now auxiliary.
Since we do not rename such variables at the moment, we need more
precise check if a variable is auxiliary or not.