NodeEliminator: Make order of nodes deterministic
Previously, we were computing vertices based on order in unordered_map
which can vary for different platforms (because of different standard
library implementation).
To unify the behaviour across platforms, we sort the vertices before we
check them.

Also, there is no need to recompute the nodes in every iteration, we
know which vertex is being removed on every iteration.