NodeEliminator: Rename auxiliary variables preemptively
Instead of trying to guess when variable renaming is necessary, and
trying to rename variables before edge merging, a better solution is
simply to rename auxialiary variables when a new edge is created (as
part of edge merging).

NOTE: We could improve auxiliary variable elimination here. Currently,
the elimination is only based on extracting substitutions, but it does
not detect, for example, that an auxiliary variable is redundant and
could be removed.