Preprocessing: Fix handling of special case in chain summarizer
Previously, the transformation would treat a loop consisting of two
edges (s1 -> s2 -> s1) as a simple chain (it would consisder vertex s2
as simple). However, this was not intended, and backtranslator does not
work correctly for this case. (I am even not sure if this transformation
is actually correct).

We fix the definition of a simple node, and leave the scenario above to
be handled by SimpleNodeEliminator.