Preprocessing: Do not try to eliminate all non-loop nodes
I believe the rationale for for this transformation was to get as close
to a transition system (or DAG of transition systems) as possible.
However, trying to eliminate all non-looping nodes can lead to an
explosion in the number of edges, and sometimes even this transformation
itself does not finish. For example, if a node has more than 1000 incoming and
outgoing edges it would have to create more than 1 000 000 edges to eliminate
such a node.