Engines: Refactor method to check implied formulas more efficiently
We have used this in Spacer when we needed to check which lemmas can be
pushed to the next level.
The idea is that we try to check if all lemmas are implied in one check.
If some is not implied, we get a counterexample model and we disable
that lemma with the help of auxiliary boolean variables.

The hope is we can avoid enough calls to the SMT solver to justify the
additional overhead.