Interpret: Fix witness handling in multi-engine run
Previously, if multiple engines finished solving at the same time, they
would both print the answer and the witness, and the output to the user
would be a mingle of everything.

We update the multi-engine run so that the child processes communicate
through a shared memory to indicate if anyone already found a solution.
We the flag is set by one child, the other ones would notice it and do
nothing. Additionally, they exit with error, because the main process
kills all children, if one finishes with exit code 0.