Fix simplification in the presence of non-var substitution keys
With arrays, it can happen that the solver finds a substitution of a
select term to, e.g., a constant.
We cannot ignore this substitution and we need to return the
corresponding equality to the constraint.
However, the term could contain some variables that were eliminated,
so we cannot simply return the equality in its original form.
In fact, we need to re-apply the substitutions with variables that we
have completely eliminated from the formula, to ensure we do not
re-introduce them (without the constraint represented by the
substitution).