KIND: Init and query must not contain auxiliary variables
K-induction checks, as implemented, require satisfiability checks where
the init and query states are negated.
If they would contain auxiliary variables, these would have to be
universally quantified in the negations.
Treating them as existentially quantified changes the meaning of the
checks. For this reason, we need to fully eliminate auxiliary variables
from Init and Query in the KIND engine.

This basically reverts the change made in 9811d3a for KIND, but keeps
it for other engines where the same issue does not apply.