Utils: Improve trivial quantifier elimination
Previously, if the code discovering substitutions would pick a variable
we do not want to eliminate as the LHS of a substitution, we would only
use this substitution if the RHS would be EXACTLY a variable we want to
eliminate.
However, we would miss a lot of cases this way where we can extract a
substitution we can use.
Here we improve this in a small way by detecting cases where RHS is a
variable shifted by a constant.

Here we only detect positive shifts (v1 = v2 + c).
We should extend it to negative shifts as well.