Symbolic execution: Prototype of forward symbolic execution on linear systems
On linear systems, the algorithm can be quite simple.
We only need to rememeber what predicate with what state (a state formula,
where non-current state variables are implicitly existentially quantified)
has been reached. Current exploration strategy is breadth-first search
where error states are checked immediately.

We need to add witness production and it might be beneficial in some
cases to try to detect if the path is infeasible (but ideally avoid a
call to the solver).