SE: Compute validity witnesses
We basically need to do what we do in PA, for each predicate collect
 all nodes in the ARG and disjoin their representations.
The problem here is that here we have representations that may contain
implicitly existentially quantified auxiliary variables.
I believe we cannot eliminate the auxiliary variables once at the end,
we need to do it for each ARG node separately. See also the comment in
the code.