PA: Check entailment of a set of formulas one by one
This was the strategy some time ago, but then it was changed between
versions 0.6.4 and 0.6.5, to unify it with the strategy used in Spacer.
However, it seems this strategy has negative effect on the performance
of PA on LRA-TS benchmarks.
We will need to search for the best strategy here, but for now we revert
to the previous implementation of the entailment check in PA.