Proofs: Update proof construction to match latest Alethe format
Since our original implementation, Alethe format has been updated.
Specifically, the `forall_inst` rule now does not take arguments in the
form of `(:= var term)`, but the arguments are just the terms that
should be used for instantiation. The connection to the variables is
implicit. The arguments are zipped with the list of quantified
variables, as given in the quantified term.

We update our implementation accordingly.
We compute the substitution as a map and then compute the correct order
of values by iterating over the quantifed variables in the quantified
term.