Proofs: Print numbers as decimals in real arithmetic
Carcara did not like that we were printing integer numbers as integers
even in real arithmetic. I believe it derives the sort from the format
of the number. Thus we need to be able to work with decimals.
There was a small issue that FastRational cannot be directly constructed
from a decimal string (GMP does not work correctly for such format).
Thus we have to work around that with OpenSMT's conversion in
ArithLogic.