Graph: Add helper to get properly versioned source terms
Previously, we were manually keeping counts for each occurence of a
predicate among the sources.
This is somewhat error-prone.
We already have a mechanism to correctly version the terms of edge
sources in the graph, so we just expose this functionality.
Here we return a vector of the source terms.
Instead of allocating a vector, we could expose a way to iterate over
the source terms an apply a given action to each.