Fix trivial quantifier elimination in the presence of arrays
Previously, we could have lost some information about the state
variables. For example, if an array access uses an auxiliary variable
that is known to be equal to a shifted state var through another
auxiliary variable.
After substitutions, the constraint is simplified to true, so all the
information is in the substitutions.
Suppose we have these substitutions:
s1[a2] -> 1
a2 -> a1
s2 -> a1 - 1

When we try to eliminate auxiliary variables a1 and a2, we figure out
that we can ignore a2, and we can eliminate a1 by reversing the last
substitution to a1 -> s2 + 1.
However, at the end, we have to bring back the equality s1[a2] = 1.
Since we brought a2 back, we have to re-apply the substitutions. But
note that just re-applying the original substitutions does not work,
since we would replace a2 with a1, but lose the information about the
state variable s2. In order to make we do not lose any constraints on
the state variables, which are on the RHS of substitutions, we have to
apply transitive closure to make sure we replace any auxiliary variables
brought back with the state variables, if they are connected.